moved some functions to new files, restructured code, changed some names for files IMPORTANT: main file is now named ws.py.
