+++ .gitignore
@@ -1,0 +1,6 @@
+# idea
+.idea/
+
+# xml/iml files
+*.xml
+*.iml
