+++ README.md
@@ -3,7 +3,7 @@
 Just trying to learn how to scrape information
 from websites and to use SQLite. This project is inspired by [Six Degrees of Wikipedia](https://github.com/jwngr/sdow).
 
-![img.png](img.png)
+![readme.png](ressources/readme.png)
 
 ## How it works
 Information is scrapped and parsed using [BeautifulSoup](https://pypi.org/project/beautifulsoup4/#:~:text=Beautiful%20Soup%20is%20a%20library,and%20modifying%20the%20parse%20tree.). The information about each page is stored as a table in an SQLite database. On user request, a database search will be done and will return a valid path if found. Otherwise, direct scraping will begin and store all the information into the database (up to 1000 pages) and will return the path if found.
